+++ src/antares/craft/service/api_services/utils.py
@@ -9,6 +9,7 @@
 # SPDX-License-Identifier: MPL-2.0
 #
 # This file is part of the Antares project.
+import io
 import time
 
 import pandas as pd
@@ -26,19 +27,13 @@ def update_series(base_url: str, study_id: str, wrapper: RequestWrapper, series:
 
 
 def get_matrix(base_url: str, study_id: str, wrapper: RequestWrapper, series_path: str) -> pd.DataFrame:
-    raw_url = f"{base_url}/studies/{study_id}/raw?path={series_path}"
+    raw_url = f"{base_url}/studies/{study_id}/raw?path={series_path}&matrix_format=arrow compressed"
     response = wrapper.get(raw_url)
-    json_df = response.json()
-
-    if "index" in json_df:
-        dataframe = pd.DataFrame(data=json_df["data"], index=json_df["index"], columns=json_df["columns"])
-    else:
-        dataframe = pd.DataFrame(data=json_df["data"], columns=json_df["columns"])
-    return dataframe
+    return pd.read_feather(io.BytesIO(response.content))
 
 
 def wait_task_completion(
-    base_url: str, wrapper: RequestWrapper, task_id: str, repeat_interval: int = 5, time_out: int = DEFAULT_TIME_OUT
+    base_url: str, wrapper: RequestWrapper, task_id: str, repeat_interval: int = 2, time_out: int = DEFAULT_TIME_OUT
 ) -> None:
     url = f"{base_url}/tasks/{task_id}"