+++ src/antares/craft/service/local_services/services/run.py
@@ -11,6 +11,7 @@
 # This file is part of the Antares project.
 import shutil
 import subprocess
+import sys
 
 from datetime import datetime
 from pathlib import Path
@@ -28,7 +29,11 @@ from antares.study.version import SolverVersion
 
 
 def _get_solver_version(solver_path: Path) -> SolverVersion:
-    process = subprocess.Popen(args=[str(solver_path) + " -v"], shell=True, stdout=subprocess.PIPE)
+    if sys.platform == "win32":
+        args = [str(solver_path), "-v"]
+    else:
+        args = [str(solver_path) + " -v"]
+    process = subprocess.Popen(args=args, shell=True, stdout=subprocess.PIPE)
     process.wait(timeout=5)
     out, _ = process.communicate()
     version_str = out.decode("utf-8").splitlines()[-1]